agents/home_assistant_agent/pyproject.toml
[tool.poetry]
name = "home-assistant-agent"
version = "0.1.0"
description = "Home Assistant Agent"
authors = ["David Snider <david@davidsnider.org>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
google-adk = "^1.11.0"
python-dotenv = "^1.0.0"
home-assistant-client = { path = "../../libs/home_assistant_client", develop = true }

[tool.poetry-monorepo-dependency-plugin]
enable = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
